requirements.txt
pandas
xlrd
pydantic
pyyaml
